- add service layer template - add default exception - add image storage service template

+++ user-service/src/main/java/ru/skillbox/userservice/model/dto/AccountDto.java
@@ -6,7 +6,7 @@ import java.time.LocalDateTime;
 
 @Data
 public class AccountDto {
-    Integer id;
+    long id;
     String email;
     String phone;
     String photo;
